third_party/WebKit/LayoutTests/platform/linux/virtual/threaded/fast/scroll-behavior/first-scroll-runs-on-compositor-expected.txt

This test verifies that the first programmatic smooth scroll doesn't always run on the main thread (see crbug.com/592799).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.getProgrammaticScrollAnimationState(document) is RUNNING_ON_COMPOSITOR
FAIL document.scrollingElement.scrollTop == 500 failed to change to true in 0.5 seconds.
PASS successfullyParsed is true

TEST COMPLETE
